search: move function to file where it's locally used
